requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
gunicorn==21.2.0
requests==2.31.0
whitenoise==6.6.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
joblib==1.3.2
